+++ sdk/appcontainers/arm-appcontainers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
     
+## 1.1.0 (2022-07-15)
+    
+**Features**
+
+  - Added operation ContainerApps.getAuthToken
+  - Added Interface ContainerAppsGetAuthTokenOptionalParams
+  - Added Type Alias ContainerAppAuthToken
+  - Added Type Alias ContainerAppsGetAuthTokenResponse
+    
+    
 ## 1.0.0 (2022-05-20)
 
 The package of @azure/arm-appcontainers is using our next generation design principles. To learn more, please refer to our documentation [Quick Start](https://aka.ms/js-track2-quickstart).